+++ zn.hpp
@@ -6,7 +6,7 @@ class ZN : public Parcelle
 {
 public:
     /* Constructors */
-    ZN(int num, string prop, Polygone<float> forme, int pConstructible = 0);
+    ZN(int num, string prop, Polygone<int> forme, int pConstructible = 0);
     ZN(Parcelle &other, int pConstructible = 0);
 
     /* Getter */
